Fixed 'Add a session' for 'UserCompany'
When adding a session for a company user, the `user_id` and `session_token` environment variables were not updated due to failed tests.

This small change fixes it by first determining if we are dealing with a regular user or a company user, before doing tests and updating Postman's env vars.